[FLINK-12240][hive] Support view related operations in GenericHiveMetastoreCatalog
